requirements.txt
psycopg2
gunicorn==20.1.0
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
whitenoise==5.0.1
requests==2.27.1